feat: update component to use default colors template and responsive
